InsightManagerForScriptrunner.groovy  * Working on splitting SimplifiedAttachmentBean to a separate class file
