+++ src/customRiadaLibraries/insightmanager/InsightManagerForScriptrunner.groovy
@@ -1114,79 +1114,6 @@ class InsightManagerForScriptrunner {
 
     }
 
-
-/**
- * <h3>This is a class intended to simplify working with Insight AttachmentBeans</h3>
- * Once instantiated it will give you easy access to the AttachmentBean it self as well as the related File object
- */
-    public class SimplifiedAttachmentBean {
-
-        public AttachmentBean attachmentBean
-        public Integer id
-        public File attachmentFile
-        public String originalFileName
-
-        SimplifiedAttachmentBean(AttachmentBean attachmentBean) {
-
-            this.attachmentBean = attachmentBean
-            this.id = attachmentBean.id
-            this.attachmentFile = getAttachmentBeanFile(this.attachmentBean)
-            this.originalFileName = attachmentBean.filename
-
-        }
-
-        boolean isValid() {
-
-            return this.attachmentBean != null && this.id > 0 && this.attachmentFile.canRead()
-
-        }
-
-        /**
-         * Compare two SimplifiedAttachmentBean to determine if they are the same
-         * @param other another SimplifiedAttachmentBean
-         * @return true if equals, false if not
-         */
-        @Override
-        boolean equals(def other) {
-            SimplifiedAttachmentBean otherBean
-
-            if (!other instanceof SimplifiedAttachmentBean) {
-                return false
-            }
-
-            otherBean = other as SimplifiedAttachmentBean
-            if (this.attachmentFile.getBytes().sha256() != otherBean.attachmentFile.getBytes().sha256()) {
-                return false
-
-            } else if (this.attachmentBean.nameInFileSystem != otherBean.attachmentBean.nameInFileSystem) {
-                return false
-            } else {
-
-                return true
-            }
-
-        }
-
-
-    }
-
-/**
- * This method will give you the File object of an AttachmentBean
- * @param attachmentBean The AttachmentBean whose File object you want
- * @return A File object
- */
-    File getAttachmentBeanFile(AttachmentBean attachmentBean) {
-        log.trace("\tGetting file for attachmentBean:" + attachmentBean.id)
-        String expectedPath = jiraDataPath + "/attachments/insight/object/${attachmentBean.objectId}/" + attachmentBean.getNameInFileSystem()
-
-
-        log.trace("\t" * 3 + "Expect file to be located here:" + expectedPath)
-
-        File attachmentFile = new File(expectedPath)
-        assert attachmentFile.canRead(): "Cant access attachment file: " + attachmentBean.getNameInFileSystem()
-        return attachmentFile
-    }
-
 /**
  * This method will retrieve all SimplifiedAttachmentBeans belonging to an object.
  * @param object key, id or objectbean
